Refactor: Create shared view model to manage new/updated collections.

+++ app/src/main/java/com/example/dndhandbook/domain/useCase/collection/deleteCollectionUseCase/DeleteCollectionUseCase.kt
@@ -7,6 +7,8 @@ class DeleteCollectionUseCase @Inject constructor(
     private val collectionRepository: CollectionContract
 ) {
     suspend fun invoke(id: Long?) {
+        if (id == -1L) return
+        
         id?.let {
             collectionRepository.getCollection(it)?.let { collection ->
                 collectionRepository.deleteCollection(collection.id)
